Changed data fetching model class , database layer class to support double data type, increased back button width height, reduced meta detail text size in detail page, removed bugs, refactored code

+++ food_ron_ai/lib/Global.dart
@@ -5,7 +5,7 @@ import 'package:food_ron_ai/stracture/ImageMetaData.dart';
 
 final appName = "Foodron.ai";
 final serve = "Serve";
-int servecount;
+double servecount;
 int counter = 1;
 final weight = "Weight";
 final calorie = "Calorie";
